backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
python-multipart
psycopg2-binary  <-- ADD THIS
python-dotenv    <-- ADD THIS (For security)